requirements.txt
openai>=1.12.0
livekit-agents>=0.1.0
livekit-plugins-deepgram>=0.1.0
livekit-plugins-openai>=0.1.0
livekit-plugins-silero>=0.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0 